requirements.txt
colour==0.1.5
coverage==4.5.3
Django==1.11
django-bootstrap3==11.0.0
django-cleanup==3.2.0
django-cors-headers==3.0.2
django-easy-select2==1.5.6
djangorestframework==3.9.4
Pillow==6.0.0
python-gettext==4.0
pytz==2019.2
django-modeltranslation==0.13.3





#python==3.7.4
#virtualenv==16.7.3